backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic==2.7.4
pydantic-settings==2.3.4
alembic==1.13.1
gunicorn==22.0.0
python-multipart==0.0.9
